hac4pro receive error fixed
git-svn-id: https://mytourbook.svn.sourceforge.net/svnroot/mytourbook/trunk@125 5a91af82-4d22-0410-8bcb-a41d7c89129f

+++ net.tourbook/src/net/tourbook/importdata/PortListener.java
@@ -240,7 +240,7 @@ public class PortListener implements SerialPortEventListener {
 				try {
 					iData = fInStream.read();
 					if (iData != -1) {
-						fDataListener.dataArrived((byte) iData);
+						fDataListener.dataArrived(iData);
 					}
 
 				} catch (IOException ex) {
